# internal/api/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server Core server.",
        "title": "GoCore Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "DuongVu",
            "url": "http://www.swagger.io/support",
            "email": "vuduongcalvin@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/login": {
            "post": {
                "description": "Logs in user by username and password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Logs in user by username and password",
                "operationId": "authLogin",
                "parameters": [
                    {
                        "description": "Credentials",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_kevinxvu_goone_internal_api_service_auth.Credentials"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/AuthToken"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/SwaggErrDetailsResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/SwaggErrDetailsResp"
                        }
                    }
                }
            }
        },
        "/refresh-token": {
            "post": {
                "description": "Refresh access token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Refresh access token",
                "operationId": "authRefreshToken",
                "parameters": [
                    {
                        "description": "RefreshTokenData",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_kevinxvu_goone_internal_api_service_auth.RefreshTokenData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/AuthToken"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/SwaggErrDetailsResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/SwaggErrDetailsResp"
                        }
                    }
                }
            }
        },
        "/v1/countries": {
            "get": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Get list country",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "countries"
                ],
                "summary": "Get list country",
                "operationId": "countriesList",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JSON string of filter. E.g: {\"field_name\":\"value\"}",
                        "name": "f",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 25,
                        "description": "Number of records per page",
                        "name": "l",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "ASC",
                            "DESC"
                        ],
                        "type": "string",
                        "default": "ASC",
                        "description": "Sort direction order must be one of ASC, DESC",
                        "name": "o",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Current page number",
                        "name": "p",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Field name for sorting",
                        "name": "s",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_kevinxvu_goone_internal_api_service_country.ListResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/SwaggErrDetailsResp"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/SwaggErrDetailsResp"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/SwaggErrDetailsResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/SwaggErrDetailsResp"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Creates new country",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "countries"
                ],
                "summary": "Creates new country",
                "operationId": "countriesCreate",
                "parameters": [
                    {
                        "description": "CreationData",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_kevinxvu_goone_internal_api_service_country.CreationData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Country"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/SwaggErrDetailsResp"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/SwaggErrDetailsResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/SwaggErrDetailsResp"
                        }
                    }
                }
            }
        },
        "/v1/countries/{id}": {
            "get": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Returns a single country",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "countries"
                ],
                "summary": "Returns a single country",
                "operationId": "countriesView",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Country ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Country"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/SwaggErrDetailsResp"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/SwaggErrDetailsResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/SwaggErrDetailsResp"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Deletes an country",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "countries"
                ],
                "summary": "Deletes an country",
                "operationId": "countriesDelete",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Country ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SwaggOKResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/SwaggErrDetailsResp"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/SwaggErrDetailsResp"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/SwaggErrDetailsResp"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/SwaggErrDetailsResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/SwaggErrDetailsResp"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Updates country information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "countries"
                ],
                "summary": "Updates country information",
                "operationId": "countriesUpdate",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Country ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "UpdateData",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_kevinxvu_goone_internal_api_service_country.UpdateData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Country"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/SwaggErrDetailsResp"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/SwaggErrDetailsResp"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/SwaggErrDetailsResp"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/SwaggErrDetailsResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/SwaggErrDetailsResp"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "AuthToken": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "expires_in": {
                    "type": "integer"
                },
                "refresh_token": {
                    "type": "string"
                },
                "token_type": {
                    "type": "string"
                }
            }
        },
        "Country": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "created_at": {
                    "description": "The time that record is created",
                    "type": "string"
                },
                "deleted_at": {
                    "description": "DeletedAt *time.Time `json:\"deleted_at\" sql:\"index\"`",
                    "allOf": [
                        {
                            "$ref": "#/definitions/gorm.DeletedAt"
                        }
                    ]
                },
                "id": {
                    "description": "ID of the record",
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "phone_code": {
                    "type": "string"
                },
                "updated_at": {
                    "description": "The latest time that record is updated",
                    "type": "string"
                }
            }
        },
        "ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "SwaggErrDetailsResp": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/ErrorResponse"
                }
            }
        },
        "SwaggOKResp": {
            "type": "object"
        },
...
        "github_com_kevinxvu_goone_internal_api_service_auth.Credentials": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "example": "superadmin123!@#"
                },
                "username": {
                    "type": "string",
                    "example": "superadmin"
                }
            }
        },
        "github_com_kevinxvu_goone_internal_api_service_auth.RefreshTokenData": {
            "type": "object",
            "required": [
                "refresh_token"
            ],
            "properties": {
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "github_com_kevinxvu_goone_internal_api_service_country.CreationData": {
            "type": "object",
            "required": [
                "code",
                "name",
                "phone_code"
            ],
            "properties": {
                "code": {
                    "description": "example: vn",
                    "type": "string",
                    "maxLength": 10,
                    "minLength": 2
                },
                "name": {
                    "description": "example: Vietnam",
                    "type": "string",
                    "minLength": 3
                },
                "phone_code": {
                    "description": "example: +84",
                    "type": "string",
                    "maxLength": 10,
                    "minLength": 2
                }
            }
        },
        "github_com_kevinxvu_goone_internal_api_service_country.ListResp": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "example: [{\"id\": 1, \"created_at\": \"2020-01-14T10:03:41Z\", \"updated_at\": \"2020-01-14T10:03:41Z\", \"name\": \"Singapore\", \"code\": \"SG\", \"phone_code\": \"+65\"}]",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Country"
                    }
                },
                "total_count": {
                    "description": "example: 1",
                    "type": "integer"
                }
            }
        },
        "github_com_kevinxvu_goone_internal_api_service_country.UpdateData": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "example: vn",
                    "type": "string",
                    "maxLength": 10,
                    "minLength": 2
                },
                "name": {
                    "description": "example: Vietnam",
                    "type": "string",
                    "minLength": 3
                },
                "phone_code": {
                    "description": "example: +84",
                    "type": "string",
                    "maxLength": 10,
                    "minLength": 2
                }
            }
        },
...
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerToken": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}